+++ EDDI/ChangeLog.md
@@ -1,6 +1,8 @@
 ﻿# CHANGE LOG
 
 ### 3.0.0-rc2
+  * Speech Responder
+    * Added 'Fighter rebuilt' event
 
 ### 3.0.0-rc1
   * Core